+++ kibot/globals.py
@@ -112,6 +112,8 @@ class Globals(FiltersOptions):
             """ Format used for the PCB and schematic date when using the file timestamp. Uses the `strftime` format """
             self.dir = ''
             """ Default pattern for the output directories """
+            self.disable_3d_alias_as_env = False
+            """ Disable the use of environment and text variables as 3D models aliases """
             self.drc_exclusions_workaround = False
             """ KiCad 6 introduced DRC exclusions. They are stored in the project but ignored by the Python API.
                 This is reported as bug number 11562 (https://gitlab.com/kicad/code/kicad/-/issues/11562).