+++ kibot/misc.py
@@ -225,6 +225,7 @@ W_NOMATCH = '(W092) '
 W_DOWNTOOL = '(W093) '
 W_NOPREFLIGHTS = '(W094) '
 W_NOPART = '(W095) '
+W_MAXDEPTH = '(W096) '
 # Somehow arbitrary, the colors are real, but can be different
 PCB_MAT_COLORS = {'fr1': "937042", 'fr2': "949d70", 'fr3': "adacb4", 'fr4': "332B16", 'fr5': "6cc290"}
 PCB_FINISH_COLORS = {'hal': "8b898c", 'hasl': "8b898c", 'imag': "8b898c", 'enig': "cfb96e", 'enepig': "cfb96e",