+++ tests/test_plot/test_step.py
@@ -78,6 +78,23 @@ def test_step_gl_env(test_dir):
 
 @pytest.mark.slow
 @pytest.mark.kicad2step
+@pytest.mark.skipif(context.ki5(), reason="uses text variables")
+def test_step_alias_1(test_dir):
+    prj = 'bom_w_prj'
+    ctx = context.TestContext(test_dir, prj, 'step_alias', STEP_DIR)
+    ctx.run(extra_debug=True)
+    # Check all outputs are there
+    name = prj+'-3D.step'
+    ctx.expect_out_file_d(name)
+    # Check the R and C 3D models are there
+    ctx.search_in_file_d(name, ['R_0805_2012Metric', 'R_0805_2012Metrico', 'C_0805_2012Metric'])
+    ctx.search_err(['Missing 3D model for R1: `(.*)R_0805_2012Metrico',
+                    'Failed to download `(.*)R_0805_2012Metrico'], invert=True)
+    ctx.clean_up(keep_project=True)
+
+
+@pytest.mark.slow
+@pytest.mark.kicad2step
 def test_step_variant_1(test_dir):
     prj = 'kibom-variant_3'
     ctx = context.TestContext(test_dir, prj, 'step_variant_1')