+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - General things:
   - Some basic preprocessing, now you can parametrize the YAML config.
     (See #233 #243)
+  - Environment and text variables now can be used as 3D model aliases.
+    (See #261)
+  - Environment and text variables expansion is now recursive.
+    So in `${VAR}` the *VAR* can contain `${OTHER_VAR}`
 - New outputs:
   - PCB_Variant: saves a PCB with filters and variants applied.
   - File_Copy: used to copy files to the output directory. (#279)